+++ examples/kalmgemma3/kalmgemma3_test.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"flag"
 	"fmt"
+	"io"
 	"math"
 	"os"
 	"runtime"
@@ -12,6 +13,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/gomlx/go-huggingface/hub"
+	"github.com/gomlx/go-huggingface/models/safetensors"
 	"github.com/gomlx/go-huggingface/models/transformer"
 	"github.com/gomlx/go-huggingface/tokenizers/api"
 	"github.com/gomlx/gomlx/backends"
@@ -29,11 +32,13 @@ import (
 var (
 	flagUseCausalMask = flag.Bool("use_causal_mask", true, "Use causal mask in the transformer: the paper suggests one shouldn't, "+
 		"but for testing it makes the result closer to Python's using HF transformer library, which seems to use it.")
-	flagListPrompts = flag.Bool("prompts", false, "During initialization lists prompts from the dataset and exit immediately.")
+	flagListPrompts        = flag.Bool("prompts", false, "During initialization lists prompts from the dataset and exit immediately.")
+	flagSkipLoadingWeights = flag.Bool("skip_loading_weights", false, "Skip loading weights from the model.")
 )
 
 var (
 	testBackend backends.Backend
+	testRepo    *hub.Repo
 	testCtx     *context.Context
 	testModel   *transformer.Model
 	taskPrompts TaskPrompts
@@ -70,13 +75,13 @@ func TestMain(m *testing.M) {
 	}
 
 	testCtx = context.New().Checked(false)
-	repo, err := LoadRepo()
+	testRepo, err = LoadRepo()
 	if err != nil {
 		fmt.Printf("Failed to LoadRepo: %v\n", err)
 		os.Exit(1)
 	}
 
-	testModel, err = transformer.LoadModel(repo)
+	testModel, err = transformer.LoadModel(testRepo)
 	if err != nil {
 		fmt.Printf("Failed to LoadModel: %v\n", err)
 		os.Exit(1)
@@ -101,7 +106,7 @@ func TestMain(m *testing.M) {
 	}
 	testPadID = int32(padID)
 
-	taskPrompts = must1(LoadTaskPrompts(repo))
+	taskPrompts = must1(LoadTaskPrompts(testRepo))
 	fmt.Printf("✅ Task prompts loaded: %d tasks\n", len(taskPrompts))
 
 	testQueries = []string{
@@ -113,13 +118,15 @@ func TestMain(m *testing.M) {
 		"Gravity is a force that attracts two bodies towards each other. It gives weight to physical objects and is responsible for the movement of planets around the sun.",
 	}
 
-	fmt.Printf("- Loading model weights ...")
-	start := time.Now()
-	must(testModel.LoadContext(testBackend, testCtx))
-	for range 3 {
-		runtime.GC()
+	if !*flagSkipLoadingWeights {
+		fmt.Printf(" - Loading model weights ...\r")
+		start := time.Now()
+		must(testModel.LoadContext(testBackend, testCtx))
+		for range 3 {
+			runtime.GC()
+		}
+		fmt.Printf("\r✅ Loading model weights: done (%v)\n", time.Since(start))
 	}
-	fmt.Printf("\r✅ Loading model weights: done (%v)\n", time.Since(start))
 
 	// Run the tests
 	code := m.Run()
@@ -566,4 +573,84 @@ func TestSimilarity(t *testing.T) {
 	fmt.Printf("- Expected: %v\n", want)
 	got := tensors.MustCopyFlatData[float32](similarities)
 	require.InDeltaSlicef(t, want, got, 1e-2, "Similaries don't match!")
+}
+
+// TestReadAllShards simply read all the shard files into /dev/null, used only
+// to test the speed.
+func TestReadAllShards(t *testing.T) {
+	var buf [1 << 20]byte
+	var f *os.File
+	defer func() {
+		if f != nil {
+			f.Close()
+		}
+	}()
+
+	fmt.Printf("- Reading all shards ...")
+	start := time.Now()
+	for filename, err := range testRepo.IterFileNames() {
+		if err != nil {
+			t.Fatalf("Failed to iterate over file names: %v", err)
+		}
+		if !strings.HasSuffix(filename, ".safetensors") {
+			continue
+		}
+		localPath := must1(testRepo.DownloadFile(filename))
+		f = must1(os.Open(localPath))
+		for {
+			_, err = f.Read(buf[:])
+			if err != nil {
+				if err == io.EOF {
+					break
+				}
+				must(err)
+			}
+		}
+
+		f.Close()
+		f = nil
+	}
+	fmt.Printf("done (%v)\n", time.Since(start))
+}
+
+func TestSafetensors(t *testing.T) {
+	if !*flagSkipLoadingWeights {
+		t.Skip("Skipping TestIterTensorsFromRepo because -skip_loading_weights flag is not set: it may not have enough accelerator space to load the model twice.")
+	}
+
+	t.Run("IterTensorsFromRepo", func(t *testing.T) {
+		start := time.Now()
+		var allTensors []safetensors.TensorAndName
+		for tan, err := range safetensors.IterTensorsFromRepo(testBackend, testRepo) {
+			require.NoError(t, err)
+			allTensors = append(allTensors, tan)
+		}
+		fmt.Printf("- %d tensors loaded in %v\n", len(allTensors), time.Since(start))
+
+		start = time.Now()
+		for _, tan := range allTensors {
+			err := tan.Tensor.FinalizeAll()
+			require.NoError(t, err)
+		}
+		fmt.Printf("- %d tensors finalized in %v\n", len(allTensors), time.Since(start))
+	})
+
+	t.Run("Model.IterTensors", func(t *testing.T) {
+		stModel, err := safetensors.New(testRepo)
+		require.NoError(t, err)
+		start := time.Now()
+		var allTensors []safetensors.TensorAndName
+		for tan, err := range stModel.IterTensors(testBackend) {
+			require.NoError(t, err)
+			allTensors = append(allTensors, tan)
+		}
+		fmt.Printf("- %d tensors loaded in %v\n", len(allTensors), time.Since(start))
+
+		start = time.Now()
+		for _, tan := range allTensors {
+			err := tan.Tensor.FinalizeAll()
+			require.NoError(t, err)
+		}
+		fmt.Printf("- %d tensors finalized in %v\n", len(allTensors), time.Since(start))
+	})
 }